support user-less URI
git-clone samples the scp-like URI syntax as:
> [user@]host.xz:path/to/repo.git/
where [] denotes an optional element. Due to the way the parse method
employed string splits this actually was not supported as optional though.

Reverse the split order on the host_part, so that iff there is no @ in the
host_part the host variable gets assigned the not split host_part and the
userinfo becomes nil (instead of having it the other way around).

Also added a test for this.